+++ CHANGELOG.md
@@ -10,12 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Sentry logging for failed livery installations (#105)
+- Added installed livery management (updating coming soon!) (#107)
 
 ### Changed
 
 - Allow access to setup stage without internet connectivity (#104)
 - Changed the "Failed to install livery" message (#105)
 - Prevent crashing when installing a corrupt livery (#105)
+- Updated dependencies (#106)
 
 ### Removed